pyproject.toml
[project]
name = "motivation-letter-agent"
version = "0.1.0"
description = "Intelligent Career Assistant with Multi-Layer Quality Control - A personalized motivation letter writer using Google ADK"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11,<3.12"
keywords = ["ai", "agents", "motivation-letter", "career", "google-adk", "multi-agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]

dependencies = [
    "google-adk>=1.18.0",
    "deprecated",
    "pandas",
    "tabulate",
    "tqdm",
    "scikit-learn",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/agent-xpeng"
Repository = "https://github.com/yourusername/agent-xpeng"
Issues = "https://github.com/yourusername/agent-xpeng/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["blogger_agent"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
